CF1-113 Add eslint-babel and ignore react prop-types
Adds the eslint settings that allow for JavaScript methods of the
following format to be linted:

```
const methodName = (args) => { body };
```

Disables the React prop-types requirement that is built-in to
babel-eslint. PropTypes seem to be type annotations that provide extra
guarantees at compile time, but may be too much boilerplate or confusing
to maintain.